refactor(discord): make recording contracts a type-only leaf
Move capture, packet-receipt, and segment-result contracts out of their
runtime owners, breaking the full import graph cycle. Narrow realtime
recording metadata to the three strings it consumes.

All six existing modules emit identical JavaScript; the new leaf emits
none. Validation: both import-cycle guards, three canonical type lanes,
18 focused tests, lint, full build, and scoped P0-P2 review passed.
